fix: improve package score page (#58)
Firstly, this fixes the algorithm to actually let you get to 100%.

Then it fixes the calculation for the symbol docs to only require 80% of
symbols to be documented.

Then it adds descriptions to the scores + a docs page for writing
documentation.

<img width="1439" alt="image"
src="https://github.com/jsr-io/jsr/assets/7829205/74760a59-7b5f-4d50-b652-4b7029f87ea6">