feat(AnimatedLogo): data-driven & random
by extracting the logo as an ASCII drawing,
and mapping over the values, we get more control
over the shape, colours, animation randomness
and positioning of the internal elements.

the asset is a snapshot of the rendered HTML
on a given page. Manual adjustments of individual
blocks is easier with inline styles.