Fix horizontal overflow for code blocks and add scrollbar colors (#138)
Some code blocks in package docs are slightly overflowing. This fixes
that, and also adds a branded color to scrollbars on some
browsers/systems (mostly so that items like these code blocks, which are
supposed to go to the edge of the page, go visibly to the edge of the
scrollbar instead).

<img width="499" alt="image"
src="https://github.com/jsr-io/jsr/assets/22334764/b87d9b1c-7659-46db-a8da-f0aab8aa577e">