feat: add jsr.json JSON schema and docs about configuring JSR (#198)
I added the json schema to the static folder under
`/schema/config-file.v1.json`, let me know if there is a better place to
put it. I've also made a docs page on configuring JSR because when I
first was looking at the JSR docs I was trying to find a page on this
but it's currently berried in publishing a package. I tried to strike a
balance between content on this new page and the publishing a package
page.

Closes #142

---------

Co-authored-by: Luca Casonato <lucacasonato@yahoo.com>
Co-authored-by: Luca Casonato <hello@lcas.dev>