feat: add sudo mode
Going forward, all admin operations on projects staff members do not normally have access to, require them to enable "sudo mode". This gives them access to all projects for the duration of enablement.

Sudo mode is enabled via the admin menu, for a duration of 5 minutes. During this time, a big red banner is displayed at the top of the screen.