feat: add sudo mode (#249)
Going forward, all admin operations on projects staff members do not
normally have access to, require them to enable "sudo mode". This gives
them access to all projects for the duration of enablement.

Sudo mode is enabled via the admin menu, for a duration of 5 minutes.
During this time, a big red banner is displayed at the top of the
screen.


No sudo:


![image](https://github.com/jsr-io/jsr/assets/7829205/b53d4727-3132-4667-879b-4312d928a610)

Enabling sudo:


![image](https://github.com/jsr-io/jsr/assets/7829205/6ce9d756-b2b1-494f-b3d5-9ff0fd813a8a)

With sudo:


![image](https://github.com/jsr-io/jsr/assets/7829205/ba8b8439-35dd-4be0-9b5f-0d9b9a8e7e6b)