fix(frontend): made sidebar on package page scroll independently (#258)
Fixes https://github.com/jsr-io/jsr/issues/41

Since the layout is designed in a way that the viewport always scrolls
(even when the page content is smaller) as it has full width top
elements with a sticky sidebar and a large whitespace before the footer
(with `class="mb-20"`) on every page, it made sense to make sidebar
scrollable independently.

If we don't want viewport to scroll when the page content is smaller, we
might want to revisit the layout. (I really like the approach of
https://docs.rs on this) But for now independent scrolling should be
sufficient imo.

---------

Co-authored-by: Luca Casonato <hello@lcas.dev>