fix: design update to Docs tab (#468)
Quite a long list of style changes affecting the Docs tab. I did them
all in the jsr repo instead of ddoc because it was an easier dev
workflow but it may be better to use this as a reference for what needs
to change in deno_doc with the goal of eliminating nearly every `body
.ddoc` style out of the stylesheet with the exception of any jsr theme
colors.

If we want to review and merge this we can but if we want to leave it as
a TODO list and then close it out when deno_doc gets the updates, that's
fine too.

also fixes #451 

![jsr
test_@oak_oak_doc](https://github.com/jsr-io/jsr/assets/776987/7d68ed6d-c19d-4701-ad84-609a52cdae4e)