fix: update styles on symbol detail pages (#476)
First attempt at updating some of the styles on the symbol detail pages.
I tried to leave some comments on bits where the thing I was changing
could actually be fixed by applying a different tailwind class in
deno_doc (like changing margin-top when I really want a `space-y-3` to
be a `space-y-2`

Same caveat as the last one, this could be used as a TODO list for
changes that should happen back in deno_doc or merged in and used as a
reference for refactoring later.

Example page of these styles on Oak/oak since that contains a large
amount of varied text to format. (sidebar omitted from screenshot since
that's changing in a different PR and I didn't want it to distract)
![jsr
test_@oak_oak_doc_~_Context](https://github.com/jsr-io/jsr/assets/776987/085d5e49-005b-4545-8452-d9f208682be2)

---------

Co-authored-by: Leo Kettmeir <crowlkats@toaxl.com>