feat: add personal access tokens (#434)
This commit adds API endpoints for managing personal access tokens.
Users
can create and revoke tokens, and tokens can be used to authenticate API
requests.

Tokens are created with a description, and an optional expiry date.
Tokens
can be revoked at any time.

There is a UI for managing tokens in the account settings - users can
create
and revoke tokens from there. Tokens are displayed with their
description,
expiry date, the granted permissions, and the date they were created.

An email is sent to users when a new token is created to notify them of
the
token's creation.

The token creation UI contains a questionnaire flow to help users
understand
that what they are doing is potentially dangerous, and to guide users to
use
alternative authentication mechanisms such as OIDC or interactive flow
where
possible.


Fixes #393