fix: make the right side of package header visible (#720)
If you try to access this package via the following URL: 

https://jsr.io/@planetarium/lib9c@0.3.0-dev.202409100854550774+2db5208d2bf73ba22624cdd8f2f7cf167097359e

you'll notice that the right side of the package header isn’t fully
visible. This happens because the version string is unusually long.

![image](https://github.com/user-attachments/assets/d8aa0bfb-3be0-4762-b827-0fcf3923f22d)

While this issue is rare, I wanted to ensure that the package details
are at least viewable. I considered implementing a more polished
solution, but it impacted other parts of the UI and required more effort
for something that doesn't occur often. So, I decided to go with a
simpler, cleaner fix that achieves the primary goal which is making the
information visible.

so now it looks like this:

![image](https://github.com/user-attachments/assets/fa57b3b9-659c-4a42-900c-8b98d793f212)