feat: include module paths in Dependencies page (#556)
Closes #310

I feel the Dependencies tab is currently quite misleading, and doesn't
reflect that only certain modules from a package are actually depended
upon.

This PR adds the individual modules paths with links to the whole
package and to the individual module docs (for npm the module path is
not linked as there is nowhere really for it to link to).

<img width="515" alt="Screenshot 2024-05-23 at 12 13 42 pm"
src="https://github.com/jsr-io/jsr/assets/71291/e7968d4d-f5f1-4ea5-a44e-bf2174b3deac">