feat: focus search input on page load (#815)
When a user opens [jsr.io](https://jsr.io), the page shows a big input
element which is used to search for packages. It appears to be the
primary call to action element and I believe by default, it should be
focused on page load because most of the times, a user would visit jsr
to search for a package.

This behaviour would also align with [google.com](https://google.com)