feat(admin): allow deleting package versions (#1011)
this PR adds an endpoint to be able to delete package versions, but only
for admins. This is not exposed in the UI for admins either, as the fact
that this is for admins only is just a temporary restriction, and the
idea is to come up with sensible restrictions that apply to users so
they can delete versions (see #899).