feat: ticketing system (#1026)
This is a first pass to the ticketing system. this requires that users
are logged in. In the future we would want to accept tickets by people
who arent logged in, but this is for a future revision.

This PR adds:
- ticket view `/ticket/:id`, which is authenticated and can only be
viewed by staff and the user that created it
<img width="1247" alt="Screenshot 2025-04-02 at 12 09 40"
src="https://github.com/user-attachments/assets/25925c00-86e7-4cfa-9ce9-8c2cbfc57246"
/>

- list of tickets for users to see their tickets `/account/tickets`
<img width="1213" alt="Screenshot 2025-04-02 at 12 18 08"
src="https://github.com/user-attachments/assets/52f52b3e-b366-4b2b-86fe-5985827f464d"
/>

- list of tickets for admins `/admin/tickets`
<img width="1232" alt="Screenshot 2025-04-02 at 12 10 10"
src="https://github.com/user-attachments/assets/f9990fe7-b022-46b9-a14d-2f255559d93d"
/>

The notifications for tickets for normal users are for their tickets,
and for staff its notifications for all tickets.
Notifications are:
- For users: any message by staff posted after the latest message by the
user
- For staff: any message by the user that was posted after the latest
message by the staff

When a non-creator of a ticket sends a message (so a staff member), an
email is sent to the user.

Tickets can have metadata associated to them (a set of key-value pairs),
and are either displayed as-is or are processed for easier
intepretation.

For the future:
- support ticket creation and messaging by non-authenticated users.
- support receiving emails for tickets (currently emails and tickets
will be managed separately). This will require supporting inbound emails
on jsr.

---------

Co-authored-by: Luca Casonato <hello@lcas.dev>