fix: use windows-compatible filename (#1070)
Cloning the JSR repository may yield errors on Windows, which seems to
stem from the `cloudtrace:v2.ts` file. This is because [semicolon may
not be a valid character in file
names](https://stackoverflow.com/a/31976060).


![image](https://github.com/user-attachments/assets/e37f35c1-6269-4f92-8cc6-37a53401a7a1)

To make it easier for Window's users to contribute without errors, I
suggest renaming `cloudtrace:v2.ts` to `cloudtrace.v2.ts`.