docs: update usage-policy.md to suggest ticket system over email (#1108)
This pull request changes the wording of the Report Violations section
to suggest using the ticket system instead of sending an email. I am not
sure if the wording is great or could be improved. There is also a few
other spots that also suggest emailing still, but there isn't a clear
button to open a ticket for it.

### PR Checklist

- [x] The PR title follows
[conventional commits](https://www.conventionalcommits.org/en/v1.0.0/)
- [ ] Is this closing an open issue? If so, link it, else include a
proper
      description of the changes and rason behind them.
- [ ] Does the PR have changes to the frontend? If so, include
screenshots or a
      recording of the changes.
<br/>If it affect colors, please include screenshots/recording in both
      light and dark mode.
- [ ] Does the PR have changes to the backend? If so, make sure tests
are added.
<br/>And if changing dababase queries, be sure you have ran `sqlx
prepare`
      and committed the changes in the `.sqlx` directory.